Models: Add UserProfile to link to Agent
Add a UserProfile extension to User for additional data, in this case linking
to an associated Agent. Fix primary key field types to what Django expects before
syncdb is run so Django can create the m2m table.

+++ src/MCPServer/share/mysql_dev_8634_events_agents.sql
@@ -8,8 +8,8 @@ ALTER TABLE Events DROP COLUMN linkingAgentIdentifier;
 -- SQL from Django
 CREATE TABLE `Events_agents` (
     `id` integer AUTO_INCREMENT NOT NULL PRIMARY KEY,
-    `event_id` bigint(20) unsigned NOT NULL,
-    `agent_id` int(10) unsigned NOT NULL,
+    `event_id` integer NOT NULL,
+    `agent_id` integer NOT NULL,
     UNIQUE (`event_id`, `agent_id`)
 )
 ;
